+++ cmd/nvidia-ctk/config/create-default/create-default.go
@@ -17,13 +17,9 @@
 package defaultsubcommand
 
 import (
-	"bytes"
 	"fmt"
-	"io"
-	"os"
-	"path/filepath"
-	"regexp"
 
+	"github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk/config/flags"
 	"github.com/NVIDIA/nvidia-container-toolkit/internal/config"
 	"github.com/NVIDIA/nvidia-container-toolkit/internal/logger"
 	"github.com/urfave/cli/v2"
@@ -31,13 +27,6 @@ import (
 
 type command struct {
 	logger logger.Interface
-}
-
-// options stores the subcommand options
-type options struct {
-	config  string
-	output  string
-	inPlace bool
 }
 
 // NewCommand constructs a default command with the specified logger
@@ -50,12 +39,12 @@ func NewCommand(logger logger.Interface) *cli.Command {
 
 // build creates the CLI command
 func (m command) build() *cli.Command {
-	opts := options{}
+	opts := flags.Options{}
 
 	// Create the 'default' command
 	c := cli.Command{
-		Name:    "generate-default",
-		Aliases: []string{"default"},
+		Name:    "default",
+		Aliases: []string{"create-default", "generate-default"},
 		Usage:   "Generate the default NVIDIA Container Toolkit configuration file",
 		Before: func(c *cli.Context) error {
 			return m.validateFlags(c, &opts)
@@ -67,117 +56,39 @@ func (m command) build() *cli.Command {
 
 	c.Flags = []cli.Flag{
 		&cli.StringFlag{
-			Name:        "config",
-			Usage:       "Specify the config file to process; The contents of this file overrides the default config",
-			Destination: &opts.config,
-		},
-		&cli.BoolFlag{
-			Name:        "in-place",
-			Aliases:     []string{"i"},
-			Usage:       "Modify the config file in-place",
-			Destination: &opts.inPlace,
-		},
-		&cli.StringFlag{
 			Name:        "output",
+			Aliases:     []string{"o"},
 			Usage:       "Specify the output file to write to; If not specified, the output is written to stdout",
-			Destination: &opts.output,
+			Destination: &opts.Output,
 		},
 	}
 
 	return &c
 }
 
-func (m command) validateFlags(c *cli.Context, opts *options) error {
-	if opts.inPlace {
-		if opts.output != "" {
-			return fmt.Errorf("cannot specify both --in-place and --output")
-		}
-		opts.output = opts.config
-	}
-	return nil
+func (m command) validateFlags(c *cli.Context, opts *flags.Options) error {
+	return opts.Validate()
 }
 
-func (m command) run(c *cli.Context, opts *options) error {
-	if err := opts.ensureOutputFolder(); err != nil {
-		return fmt.Errorf("unable to create output directory: %v", err)
-	}
-
-	contents, err := opts.getFormattedConfig()
+func (m command) run(c *cli.Context, opts *flags.Options) error {
+	cfgToml, err := config.New()
 	if err != nil {
-		return fmt.Errorf("unable to fix comments: %v", err)
+		return fmt.Errorf("unable to load or create config: %v", err)
 	}
 
-	if _, err := opts.Write(contents); err != nil {
-		return fmt.Errorf("unable to write to output: %v", err)
+	if err := opts.EnsureOutputFolder(); err != nil {
+		return fmt.Errorf("failed to create output directory: %v", err)
+	}
+	output, err := opts.CreateOutput()
+	if err != nil {
+		return fmt.Errorf("failed to open output file: %v", err)
+	}
+	defer output.Close()
+
+	_, err = cfgToml.Save(output)
+	if err != nil {
+		return fmt.Errorf("failed to write output: %v", err)
 	}
 
 	return nil
-}
-
-// getFormattedConfig returns the default config formatted as required from the specified config file.
-// The config is then formatted as required.
-// No indentation is used and comments are modified so that there is no space
-// after the '#' character.
-func (opts options) getFormattedConfig() ([]byte, error) {
-	cfg, err := config.Load(opts.config)
-	if err != nil {
-		return nil, fmt.Errorf("unable to load or create config: %v", err)
-	}
-
-	buffer := bytes.NewBuffer(nil)
-
-	if _, err := cfg.Save(buffer); err != nil {
-		return nil, fmt.Errorf("unable to save config: %v", err)
-	}
-	return fixComments(buffer.Bytes())
-}
-
-func fixComments(contents []byte) ([]byte, error) {
-	r, err := regexp.Compile(`(\n*)\s*?#\s*(\S.*)`)
-	if err != nil {
-		return nil, fmt.Errorf("unable to compile regexp: %v", err)
-	}
-	replaced := r.ReplaceAll(contents, []byte("$1#$2"))
-
-	return replaced, nil
-}
-
-func (opts options) outputExists() (bool, error) {
-	if opts.output == "" {
-		return false, nil
-	}
-	_, err := os.Stat(opts.output)
-	if err == nil {
-		return true, nil
-	} else if !os.IsNotExist(err) {
-		return false, fmt.Errorf("unable to stat output file: %v", err)
-	}
-	return false, nil
-}
-
-func (opts options) ensureOutputFolder() error {
-	if opts.output == "" {
-		return nil
-	}
-	if dir := filepath.Dir(opts.output); dir != "" {
-		return os.MkdirAll(dir, 0755)
-	}
-	return nil
-}
-
-// Write writes the contents to the output file specified in the options.
-func (opts options) Write(contents []byte) (int, error) {
-	var output io.Writer
-	if opts.output == "" {
-		output = os.Stdout
-	} else {
-		outputFile, err := os.Create(opts.output)
-		if err != nil {
-			return 0, fmt.Errorf("unable to create output file: %v", err)
-		}
-		defer outputFile.Close()
-		output = outputFile
-	}
-
-	return output.Write(contents)
 }